requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-barcode==0.15.1
pillow==11.3.0
segno==1.6.6
treepoem==3.27.1